firmware: use PF instead of EF/FF and INFM1/OEP1. "NFCI."
This allows us greater freedom in changing the latency of the FX2
arbiter for IN and OUT transfers alike.

This is a fairly significant change that could lead to corruption of
data in transit and hangs, but it seems OK based on testing with
the following applets:
  * benchmark
  * memory-25x
  * video-rgb-input

Also, gruetzkopf has verified this according to documentation and
silicon behavior and it also appears fully equivalent.