access.direct.demultiplexer: pipeline reads and writes.
This significantly improves throughput (up to ~21 MB/s) without any
particular action on behalf of applets.

Hinting is now obsolete. Two applets used hinting:
  * memory-floppy: hinting is removed (untested);
  * audio-yamaha-opl: hinting and custom pipelining are removed.

The audio-yamaha-opl applet requires explicit flushes now, since
otherwise the playback coroutine will not yield nearly often enough.

The `benchmark loopback` mode also now works, since it was hanging
not because of the arbiter (the arbiter will not, in fact, cause
starvation) but because of a deadlock between read and write
coroutines inherent in the old demultiplexer implementation.

Fixes #73.