applet.memory.onfi: mark as preview-quality again.
The applet works decently enough, but the parallel bus should really
use IOB registers. It is not completely clear but this might be
the cause of read errors that manifest as single bit flips.

The reason I suspect these errors aren't inherent to the memory
itself is because they seem to correlate with the wire length in
my wiring harness. See #116 for details.

In principle, these single bit flips aren't too bad as the flash is
supposed to be read with ECC anyway, but we don't implement any ECC,
and moreover I don't feel comfortable signing off on an applet with
known data corruption issues. (It'd be much worse if the bit flip
happened in the command or address phase, too.)

Also, this applet misses ONFI signature readout (since I can't find
any ONFI-compliant flash) and/or some other way to detect page size
(yet to be discovered), which is a serious usability impediment.
Though it's less bad than the bit flips.