gateware.fx2_crossbar: rework to handle pipelining. (UPDATE FIRMWARE)
This commit reverts all PF changes and avoids relying on FF to handle
IN FIFO. Instead, the level of FX2-side FIFO is now mirrored on
the FPGA side, and PKTEND is asserted at precisely the right cycle.

Moreover, ZLPs are now sent when appropriate to avoid apparent hangs
when issuing BULK IN URBs that are larger than one packet size.

This commit doesn't yet correctly re-register the signals that are
FX2 outputs, but it's very close.

Overall, the crossbar is significantly simpler, shorter, faster and
more reliable now, and the firmware is smaller.