applet.memory.floppy: change pinout to allow straightforward crimp.
This also adds support for both pins in DRVSA/B and MOTEA/B pairs.
Currently both DRVSx pins are enabled, which allows connecting
a single floppy no matter how it is keyed. This might be changed
in the future in the unlikely case that someone wants to connect
two floppy drives to one Glasgow.

REDWC is essentially omitted as we don't (directly) support either
5.25" drives (no reason other than me not having one) or any writes
at all (not implemented yet).