access.direct.demultiplexer: add a workaround for Windows brokenness.
Yes, "brokenness". KMDF not allowing to select any configuration
other than the first one is a genuine design fault and it needs to
be fixed.

In the future we probably should add a second configuration that
allows selecting 2-FIFO or 4-FIFO mode in parallel to the standard
Set Configuration request, unless there is an UMDF libusb somewhere
(I don't think there is). That's pretty gross though.