applet.internal.selftest: somewhat reduce false positives.
Ideally this would output the results in terms of measured tRC,
because marginal failures appear stochastic in terms of PASS/FAIL.

Also, reword the output to use "fail" instead of "stuck" because e.g.
shifter input buffer degrading such that it has a weak pullup to
Vio is a "fail high" but it is not "stuck high" because the failure
is not well repeatable.