software: drop Python 3.6 support (for real this time).
This was done about half a year ago and then reverted in ae86b060 by
popular request. The main reason to migrate is the proper support for
required subparser arguments in argparse, the lack of which caused
rather frustrating incidents: forgetting to specify a command would
silently do nothing, which without verbose logging enabled can look
identical to a command executing successfully.