applet.memory.prom: encourage (sic) metastability during reads.
This change is tested with Atmel AT29C flash and may or may not have
any effect on other memories. In case of this Atmel flash, output
buffers have parasitic feedback (likely through power supply) and
will oscillate if encouraged, e.g. if OE is kept asserted, or if data
transitions are maximized.