revC2: fix distance between port headers to multiple of 2.54 mm
It was off by 0.1 mm before, thanks to user disasm for spotting this.

A raster of 2.54 mm is desirable for compatibility to other 2.54 mm
equipment like protoboards. Also designing addon boards is simplified
by allowing to use an imperial grid in CAD.

Also update the dimension drawings. Align the dimension drawings to
the location of pin 1 instead of the connector center, as Kicad library
connectors have their origin on pin 1 and not at the center.