revC2: Addressed some DRC issues.
Addressed:
* Decreased via size under the BGA. 0.2mm drill 0.4mm pad, this
increases the clearance to the pads and is still a reasonable
requirement for most PCB houses. (PCB houses can adjust these if
necessary in most cases)
* Decreased via size under the ESD diode packs.

Disabled:
* Disabled courtyard checking, this is a guidance in the first place, also
we violate it in so many places that the DRC errors drown in the noise.
It is still good practice to respect courtyards when possible.

Not resolved:
* Pad to drill distance for the LVDS connector. To resolve it we can
either use complex pads or use the new single corner chamfered pads in
KiCad nightly. To save time I leave this issue for the very last step of
gerber generation that can be done in KiCad nightly.