firmware: move ICE LED handling to FPGA code. NFCI.
The existence of an FX2 pin dedicated to the ICE LED is an oversight;
it should have been hooked up directly to CDONE to save a scarce FX2
pin. Handle the LED transparently in FPGA code, with the aim of
eventually freeing its pin on a future revision.

Also, eliminate redundant state in the form of ST_FPGA_RDY latched
status bit.