firmware: remove LED abstraction. NFCI.
This abstraction never worked well; the ACT LED manipulation was half
inlined because function calls are tricky in interrupts, and the ICE
LED manipulation was moved to the FPGA code in commit 155244f9. Get
rid of it, simplify the code, and make the firmware smaller.

Also, rename LED constants to match the board silkscreen.