firmware: switch the DAC off too when switching off Vio
Without this, the DAC still outputs it's control voltage even if Vio is off.
This goes through the feedback resistors to the output. I measured about
15 mV without this patch on a port that was previously enabled.
With this patch it properly goes down to zero.

While such a low voltage usually shouldn't be a problem, I consider it
unclean design, especially if it is so easy to fix.