revC3: connect the 3v3 pin of the LVDS connector with a second via
Power traces should be connected with at least two vias for lower impedance
and improved reliability. On the other instances this was already addressed
in revC2, this one was missed.

Without this improvement the planned Hyperram addon for the LVDS connector
could get a power problem because 2 active Hyperram ICs draw in the order
of 100 mA through the 3v3 pin.