applet: avoid UnusedElaboratable warning spam in hardware testcases
The hardware tests using `applet_hardware_test` work by constructing the
whole applet (including the gateware), then discarding the gateware and
connecting the applet software to a mock interface. This results in a
lot of `UnusedElaboratable` warnings in CI logs, making them effectively
unusable.

This patch performs a dummy elaboration of the hierarchy to mark all
involved elaboratables as used.