firmware,software: set high byte of bcdDevice to API level.
Before this commit, bcdDevice can have two possible values:
* 0x00RR for devices without firmware
* 0x01RR for devices with firmware
(where RR is the encoded revision, e.g. 0x32 for revC2)
Finding out the API level requires opening a device and sending
the REQ_USB_API_LEVEL control request.
The API level is also 1 at the moment.

After this commit, the high byte of bcdDevice is either zero (if
the firmware is not loaded), or equal to the current API level (if
the firmware is loaded). Conveniently, with the API level before
this commit the rules are already satisfied.
REQ_USB_API_LEVEL is eliminated and finding out the API level can
now be done without opening the device.

The API level is increased to 2.

Compatibility considerations:
* The encoded hardware revision can be 0x00 in the FX2 API block
  in case of really old devices. The firmware will fix it up on load.
* There is a range of firmware between the commit that set high byte
  of bcdDevice to 0x01 (3b9b39bc29) and the commit that introduced
  API levels (d07cf5c104) where the firmware may not actually comply
  with API level 1. Because of this, this commit bumps the API level.
* A device with no FX2 configuraiton block will have bcdDevice set
  to 0xA000. (Likely, FX2 die revision A.) However, this is not
  relevant since the VID/PID are different.
* In the future, REQ_USB_API_LEVEL can be removed. However, it is
  left in for now, for compatibility with old software checkouts.