firmware: remove config fixup functionality.
There was only a small amount of devices that required config fixup
in first place, and they should have all been fixed in the ~4 years
that have passed since the code was added. If anyone boots firmware
on a device with non-fixed EEPROM after this commit, the device will
come up with a `Z9-9999...9` serial and running `glasgow flash` with
the right revision will fix the problem.

Fixes #387.